Fix path encoding
Databases might have any character in them as long as they are URL
encoded. Unfortunately this breaks when using Golang url.Parse because
it will decode the path replacing %2F with / as per
https://github.com/golang/go/commit/874a605af0764a8f340c3de65406963f514e21bc.
Making use of url.RawPath makes sure we fix this problem by the provided
workaround and trust what the user is passing to us.